requirements-colab.txt
tensorboardX
pillow
pillow-simd
h5py-cache
configargparse
sacrebleu
rouge-score
datasets==2.4.0
transformers==4.21.3
nltk
cached_property
networkx
penman>=1.1.0
pytorch-ignite
regex
smatch
wandb
amrlib
PyYAML>=5.1
sentencepiece
sacremoses
sentence-transformers
evaluate
scylla-driver
python-dotenv
